API: change focus policy + rm nb special case
Do not focus on creation, user must actively click on the figure
to get focus (and hence have the figure keyboard events).

This allows us to remove the shift-enter special case (which will be
configurable in jh).

Disable propagation on any keyboard events that the figure gets, which
removes the need for the keyboard_manager special casing.