Make handle_binary compliant with the KernelMessage.IMessage spec
Since Jupyterlab v4.0.0a21 the buffers deserialized from a KernelMessage
are now of type ArrayBuffer instead of the DataView type
handle_binary expects. However, it turns out that the spec for
KernelMessage.IMessage has always specifed that buffers could be
either an ArrayBuffer or an ArrayBufferView. This commit ensures
either type is accepted.